元に戻した

+++ EatingBuild/Assets/GameManager.cs
@@ -8,7 +8,7 @@ using DG.Tweening;
 public class GameManager : MonoBehaviour
 {
 
-    public float point;//大きさを変える時などに使うポイント
+    public int point;//大きさを変える時などに使うポイント
     [SerializeField] private EatObjectScript playerEat;
     [SerializeField] private Text scoreText;
     [SerializeField] private TextMeshProUGUI addScoreText;//+1 +2 と画面に何ポイント追加したか表示する
@@ -22,7 +22,7 @@ public class GameManager : MonoBehaviour
 
 
 
-    public void AddPoint(float number)//ポイントの追加
+    public void AddPoint(int number)//ポイントの追加
     {
         point = point + number;
         addScoreText.text = $"+{number}";
